refactor magpie init config into magpie app (single location) + apply adjustments to play nice with more recent pyramid_retry==2.1.1
