+++ requirements.txt
@@ -31,7 +31,7 @@ pyramid>=1.10.2,<2
 pyramid_beaker==0.8
 pyramid_chameleon>=0.3
 pyramid_mako>=1.0.2
-pyramid_retry==1.0
+pyramid_retry==2.1.1
 pyramid_tm>=2.2.1
 python-dotenv
 python2-secrets; python_version <= "3.5"
